requirements.txt
Flask>=2.0.0
python-dotenv>=0.15.0
Flask-SQLAlchemy>=2.5.0
Flask-Migrate>=3.0.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
meraki>=1.0.0
